move endpoint to /items/{item_id}/reservations route
